requirements.txt
# Bild-Pipeline, GUIs (tkinter/dearpygui/gradio), Excel-Listen
dearpygui>=1.10
gradio>=4.0
loguru>=0.7
numpy>=1.24
opencv-python>=4.8
openpyxl>=3.1
pandas>=2.0
pillow>=10.0
rembg>=2.0
requests>=2.31
